+++ tempest_fastapi_sdk/__init__.py
@@ -434,7 +434,7 @@ from tempest_fastapi_sdk.websockets import (
     make_websocket_router,
 )
 
-__version__: str = "0.161.0"
+__version__: str = "0.162.0"
 
 __all__: list[str] = [
     "BASE_COLUMN_ORDER",